removed zone parallel mode in noise checking
