+++ src/mqt/yaqs/simulator.py
@@ -136,7 +136,6 @@ from .core.parallel_utils import (
 )
 from .core.random_utils import make_disorder_rng, make_trajectory_rng
 from .digital.digital_tjm import digital_tjm
-from .digital.stochastic_circuit import _validate_pauli_noise_model, sample_stochastic_circuit
 from .digital.utils.qasm_utils import load_circuit
 
 __all__ = ["Simulator", "available_cpus"]
@@ -211,20 +210,6 @@ def _lindblad_ctx_worker(_traj_idx: int) -> tuple[NDArray[np.float64], None, NDA
     return lindblad_evolve(WORKER_CTX["ctx"])
 
 
-def _sample_and_run_stochastic_trajectory(
-    args: tuple[int, MPS, NoiseModel | None, DigitalSimParams, QuantumCircuit],
-) -> tuple[NDArray[np.float64] | None, NDArray[np.float64] | None, dict[int, int] | None, MPS | None]:
-    """Sample and execute one unaggregated stochastic-circuit trajectory.
-
-    Returns:
-        Raw observable data, diagnostics, counts, and optional final state.
-    """
-    trajectory_index, initial_state, noise_model, sim_params, circuit = args
-    rng = make_trajectory_rng(trajectory_index, base_seed=sim_params.random_seed)
-    noisy_circuit = sample_stochastic_circuit(circuit, noise_model or NoiseModel(), rng)
-    return digital_tjm((trajectory_index, initial_state, None, sim_params, noisy_circuit), rng=rng)
-
-
 def _digital_worker(
     traj_idx: int,
 ) -> tuple[NDArray[np.float64] | None, NDArray[np.float64] | None, dict[int, int] | None, MPS | None]:
@@ -233,16 +218,13 @@ def _digital_worker(
     Returns:
         Observable data, diagnostics, shot counts, and optional final MPS (any may be ``None``).
     """
-    args = (
+    return digital_tjm((
         traj_idx,
         WORKER_CTX["initial_state"],
         WORKER_CTX["noise_model"],
         WORKER_CTX["sim_params"],
         WORKER_CTX["operator"],
-    )
-    if WORKER_CTX.get("stochastic_circuit", False):
-        return _sample_and_run_stochastic_trajectory(args)
-    return digital_tjm(args)
+    ))
 
 
 _ProgramSegmentTrajectory = tuple[
@@ -961,60 +943,6 @@ class Simulator:
     # -----------------------------------------------------------------------
     # Public API
     # -----------------------------------------------------------------------
-    def run_stochastic_circuit(
-        self,
-        initial_state: State,
-        circuit: QuantumCircuit,
-        sim_params: DigitalSimParams,
-        noise_model: NoiseModel | None = None,
-    ) -> Result:
-        """Run an ensemble of explicit stochastic Pauli circuit realizations.
-
-        Distribution-valued strengths are resolved once per run. Each trajectory
-        independently samples one complete noisy circuit, executes it once with
-        no runtime noise model, and uses the standard digital result aggregation.
-
-        Args:
-            initial_state: Initial MPS-backed state.
-            circuit: Circuit to execute.
-            sim_params: Digital simulation parameters.
-            noise_model: Existing YAQS noise model containing Pauli processes.
-
-        Returns:
-            Aggregated simulation result.
-
-        Raises:
-            TypeError: If the inputs are not a State, QuantumCircuit, and DigitalSimParams.
-            ValueError: If no output is requested or positive-rate non-Pauli noise is supplied.
-        """
-        if not isinstance(initial_state, State):
-            msg = "Circuit simulation requires a State initial_state."
-            raise TypeError(msg)
-        if not isinstance(circuit, QuantumCircuit):
-            msg = "Circuit simulation requires a QuantumCircuit operator."
-            raise TypeError(msg)
-        if not isinstance(sim_params, DigitalSimParams):
-            msg = f"sim_params must be DigitalSimParams, got {type(sim_params).__name__}."
-            raise TypeError(msg)
-        if not sim_params.get_state and not sim_params.observables and sim_params.shots is None:
-            msg = "No output specified: set observables, shots, and/or get_state."
-            raise ValueError(msg)
-
-        if noise_model is not None:
-            noise_model = noise_model.sample(rng=make_disorder_rng(base_seed=sim_params.random_seed))
-            _validate_pauli_noise_model(noise_model)
-
-        result = Result(sim_params=sim_params, noise_model=noise_model)
-        self._run_circuit(
-            initial_state,
-            circuit,
-            sim_params,
-            noise_model,
-            result,
-            stochastic_circuit=True,
-        )
-        return result
-
     def run(
         self,
         initial_state: State | list[State],
@@ -1508,8 +1436,6 @@ class Simulator:
         sim_params: DigitalSimParams,
         noise_model: NoiseModel | None,
         result: Result,
-        *,
-        stochastic_circuit: bool = False,
     ) -> None:
         """Run digital circuit simulation trajectories.
 
@@ -1526,11 +1452,6 @@ class Simulator:
         Raises:
             ValueError: If ``get_state`` is ``True`` with a non-trivial noise model.
         """
-        backend: Callable[
-            [tuple[int, MPS, NoiseModel | None, DigitalSimParams, QuantumCircuit]],
-            Any,
-        ] = _sample_and_run_stochastic_trajectory if stochastic_circuit else digital_tjm
-
         wants_obs = bool(sim_params.observables)
         wants_shots = sim_params.shots is not None
         shots_only = wants_shots and not wants_obs
@@ -1583,7 +1504,6 @@ class Simulator:
             "noise_model": noise_model,
             "sim_params": worker_params,
             "operator": operator,
-            "stochastic_circuit": stochastic_circuit,
         }
         if per_call_shots is not None:
             payload["per_call_shots"] = per_call_shots
@@ -1642,7 +1562,7 @@ class Simulator:
                 iterator = tqdm(args, desc="Running trajectories", ncols=80, disable=not self.show_progress)
                 for i, arg in enumerate(iterator):
                     traj_data, traj_diag, shot_counts, traj_final = call_serial_capped(
-                        backend, arg, n_threads=n_threads
+                        digital_tjm, arg, n_threads=n_threads
                     )
                     _consume(i, traj_data, traj_diag, shot_counts, traj_final)
         finally:
@@ -1664,8 +1584,6 @@ class Simulator:
         sim_params: DigitalSimParams,
         noise_model: NoiseModel | None,
         result: Result,
-        *,
-        stochastic_circuit: bool = False,
     ) -> None:
         """Run circuit-based simulation trajectories.
 
@@ -1703,7 +1621,6 @@ class Simulator:
             sim_params,
             noise_model,
             result,
-            stochastic_circuit=stochastic_circuit,
         )
 
     # -----------------------------------------------------------------------
